refactor: 모든 플레이어 승패 결과 반환 로직 이동
