+++ src/test/java/domain/game/PlayersTest.java
@@ -3,8 +3,12 @@ package domain.game;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import domain.card.Card;
 import domain.card.CardDeck;
+import domain.card.CardNumber;
 import domain.card.CardShuffler;
+import domain.card.Pattern;
+import domain.card.TestShuffler;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -77,5 +81,148 @@ public class PlayersTest {
 
         //then
         Assertions.assertThat(allPlayerNames).containsExactly("aa", "bb");
+    }
+
+    @Test
+    void 모든_플레이어들과_딜러_사이의_승패를_판정한다() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer(cardDeck);
+        Hand dealerHand = dealer.getHand();
+        List<Card> dealerCards = dealerHand.getCards();
+        dealerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        dealerCards.add(new Card(Pattern.CLOVER, CardNumber.SEVEN));
+
+        Players players = new Players(List.of("pobi"), cardDeck);
+        Hand playerHand = players.getPlayers().getFirst().getHand();
+        List<Card> playerCards = playerHand.getCards();
+        playerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        playerCards.add(new Card(Pattern.CLOVER, CardNumber.SIX));
+
+        //when
+        List<GameResult> gameResult = players.judgeGameResult(dealer);
+
+        //then
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
+    }
+
+    @Test
+    void 모든_플레이어들과_딜러_사이의_승패를_판정한다_둘다_버스트() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer(cardDeck);
+        Hand dealerHand = dealer.getHand();
+        List<Card> dealerCards = dealerHand.getCards();
+        dealerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        dealerCards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
+        dealerCards.add(new Card(Pattern.HEART, CardNumber.TEN));
+
+        Players players = new Players(List.of("pobi"), cardDeck);
+        Hand playerHand = players.getPlayers().getFirst().getHand();
+        List<Card> playerCards = playerHand.getCards();
+        playerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        playerCards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
+        playerCards.add(new Card(Pattern.HEART, CardNumber.TEN));
+
+        //when
+        List<GameResult> gameResult = players.judgeGameResult(dealer);
+
+        //then
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
+    }
+
+    @Test
+    void 모든_플레이어들과_딜러_사이의_승패를_판정한다_딜러만_버스트() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer(cardDeck);
+        Hand dealerHand = dealer.getHand();
+        List<Card> dealerCards = dealerHand.getCards();
+        dealerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        dealerCards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
+        dealerCards.add(new Card(Pattern.HEART, CardNumber.TEN));
+
+        Players players = new Players(List.of("pobi"), cardDeck);
+        Hand playerHand = players.getPlayers().getFirst().getHand();
+        List<Card> playerCards = playerHand.getCards();
+        playerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        playerCards.add(new Card(Pattern.CLOVER, CardNumber.TEN));
+
+        //when
+        List<GameResult> gameResult = players.judgeGameResult(dealer);
+
+        //then
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.WIN));
+    }
+
+    @Test
+    void 모든_플레이어들과_딜러_사이의_승패를_판정한다_둘다_블랙잭() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer(cardDeck);
+        Hand dealerHand = dealer.getHand();
+        List<Card> dealerCards = dealerHand.getCards();
+        dealerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        dealerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
+
+        Players players = new Players(List.of("pobi"), cardDeck);
+        Hand playerHand = players.getPlayers().getFirst().getHand();
+        List<Card> playerCards = playerHand.getCards();
+        playerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
+
+        //when
+        List<GameResult> gameResult = players.judgeGameResult(dealer);
+
+        //then
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.DRAW));
+    }
+
+    @Test
+    void 모든_플레이어들과_딜러_사이의_승패를_판정한다_딜러만_블랙잭() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer(cardDeck);
+        Hand dealerHand = dealer.getHand();
+        List<Card> dealerCards = dealerHand.getCards();
+        dealerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        dealerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
+
+        Players players = new Players(List.of("pobi"), cardDeck);
+        Hand playerHand = players.getPlayers().getFirst().getHand();
+        List<Card> playerCards = playerHand.getCards();
+        playerCards.add(new Card(Pattern.SPADE, CardNumber.EIGHT));
+        playerCards.add(new Card(Pattern.SPADE, CardNumber.TWO));
+        playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
+
+        //when
+        List<GameResult> gameResult = players.judgeGameResult(dealer);
+
+        //then
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.LOSE));
+    }
+
+    @Test
+    void 모든_플레이어들과_딜러_사이의_승패를_판정한다_플레이어만_블랙잭() {
+        //given
+        CardDeck cardDeck = CardDeck.createCards(new TestShuffler());
+        Dealer dealer = new Dealer(cardDeck);
+        Hand dealerHand = dealer.getHand();
+        List<Card> dealerCards = dealerHand.getCards();
+        dealerCards.add(new Card(Pattern.SPADE, CardNumber.EIGHT));
+        dealerCards.add(new Card(Pattern.SPADE, CardNumber.TWO));
+        dealerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
+
+        Players players = new Players(List.of("pobi"), cardDeck);
+        Hand playerHand = players.getPlayers().getFirst().getHand();
+        List<Card> playerCards = playerHand.getCards();
+        playerCards.add(new Card(Pattern.SPADE, CardNumber.TEN));
+        playerCards.add(new Card(Pattern.CLOVER, CardNumber.ACE));
+
+        //when
+        List<GameResult> gameResult = players.judgeGameResult(dealer);
+
+        //then
+        Assertions.assertThat(gameResult).containsExactlyElementsOf(List.of(GameResult.WIN));
     }
 }
